callosum-simulator/requirements.txt
numpy==1.21
pandas
pygame
soundfile
python-dotenv
google-cloud-speech
librosa
numba
spotipy
youtube_dl
pydub
lyricsgenius
nltk
eventlet
flask
flask-cors
flask-socketio
tqdm
pyserial
